[GRAD2-1409] Update security implementation of all APIs migrated from springboot 2.4 to 2.7
