+++ api/src/main/java/ca/bc/gov/educ/api/ruleengine/EducRuleEngineApiApplication.java
@@ -1,22 +1,16 @@
 package ca.bc.gov.educ.api.ruleengine;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
 import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication(exclude = {
         DataSourceAutoConfiguration.class, 
         DataSourceTransactionManagerAutoConfiguration.class,
         HibernateJpaAutoConfiguration.class})
-@EnableGlobalMethodSecurity(prePostEnabled = true)
 @EnableCaching
 public class EducRuleEngineApiApplication {
 
@@ -24,13 +18,5 @@ public class EducRuleEngineApiApplication {
 		SpringApplication.run(EducRuleEngineApiApplication.class, args);
 	}
 
-	@Bean
-	public ModelMapper modelMapper() {
-		return new ModelMapper();
-	}
-	
-	@Bean
-	public RestTemplate restTemplate(RestTemplateBuilder builder) {
-		return builder.build();
-	}
+
 }
