+++ api/src/main/java/ca/bc/gov/educ/api/ruleengine/config/GradRuleEngineApiConfig.java
@@ -1,0 +1,22 @@
+package ca.bc.gov.educ.api.ruleengine.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class GradRuleEngineApiConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+
+}
